Warnings removed

+++ Kernel/irqDispatcher.c
@@ -100,7 +100,7 @@ uint64_t int_80(uint64_t rdi, uint64_t rsi, uint64_t rdx, uint64_t rcx, uint64_t
 			sys_printRegisters();
 			break;
 		case 20:
-			return sys_malloc((uint32_t) rsi);
+			return (uint64_t)sys_malloc((uint32_t) rsi);
 			break;
 		case 21:
 			sys_free((void *) rsi);
